+++ base/process.jl
@@ -47,40 +47,19 @@ function show(io::IO, cmd::Cmd)
     (print_dir || print_env) && print(io, ")")
 end
 
-function show(io::IO, cmds::OrCmds)
-    if isa(cmds.a, AndCmds) || isa(cmds.a, CmdRedirect)
-        print(io, "(")
-        show(io, cmds.a)
-        print(io, ")")
-    else
-        show(io, cmds.a)
-    end
-    print(io, " |> ")
-    if isa(cmds.b, AndCmds) || isa(cmds.b, CmdRedirect)
-        print(io, "(")
-        show(io, cmds.b)
-        print(io, ")")
-    else
-        show(io, cmds.b)
-    end
+function show(io::IO, cmds::Union(OrCmds,ErrOrCmds))
+    print(io, "pipe(")
+    show(io, cmds.a)
+    print(io, ", ")
+    print(io, isa(cmds, ErrOrCmds) ? "stderr=" : "stdout=")
+    show(io, cmds.b)
+    print(io, ")")
 end
 
 function show(io::IO, cmds::AndCmds)
-    if isa(cmds.a, OrCmds) || isa(cmds.a, CmdRedirect)
-        print(io, "(")
-        show(io, cmds.a)
-        print(io, ")")
-    else
-        show(io, cmds.a)
-    end
+    show(io, cmds.a)
     print(io, " & ")
-    if isa(cmds.b, OrCmds) || isa(cmds.b, CmdRedirect)
-        print(io, "(")
-        show(io, cmds.b)
-        print(io, ")")
-    else
-        show(io, cmds.b)
-    end
+    show(io, cmds.b)
 end
 
 const STDIN_NO  = 0
@@ -120,19 +99,18 @@ type CmdRedirect <: AbstractCmd
 end
 
 function show(io::IO, cr::CmdRedirect)
+    print(io, "pipe(")
+    show(io, cr.cmd)
+    print(io, ", ")
     if cr.stream_no == STDOUT_NO
-        show(io, cr.cmd)
-        print(io, " |> ")
-        show(io, cr.handle)
+        print(io, "stdout=")
     elseif cr.stream_no == STDERR_NO
-        show(io, cr.cmd)
-        print(io, " .> ")
-        show(io, cr.handle)
+        print(io, "stderr=")
     elseif cr.stream_no == STDIN_NO
-        show(io, cr.handle)
-        print(io, " |> ")
-        show(io, cr.cmd)
+        print(io, "stdin=")
     end
+    show(io, cr.handle)
+    print(io, ")")
 end
 
 
@@ -145,21 +123,41 @@ setenv(cmd::Cmd, env::Associative; dir="") = (cmd.env = ByteString[string(k)*"="
 setenv(cmd::Cmd; dir="") = (cmd.dir = dir; cmd)
 
 (&)(left::AbstractCmd, right::AbstractCmd) = AndCmds(left, right)
-(|>)(src::AbstractCmd, dest::AbstractCmd) = OrCmds(src, dest)
-(.>)(src::AbstractCmd, dest::AbstractCmd) = ErrOrCmds(src, dest)
+redir_out(src::AbstractCmd, dest::AbstractCmd) = OrCmds(src, dest)
+redir_err(src::AbstractCmd, dest::AbstractCmd) = ErrOrCmds(src, dest)
 
 # Stream Redirects
-(|>)(dest::Redirectable, src::AbstractCmd) = CmdRedirect(src, dest, STDIN_NO)
-(|>)(src::AbstractCmd, dest::Redirectable) = CmdRedirect(src, dest, STDOUT_NO)
-(.>)(src::AbstractCmd, dest::Redirectable) = CmdRedirect(src, dest, STDERR_NO)
+redir_out(dest::Redirectable, src::AbstractCmd) = CmdRedirect(src, dest, STDIN_NO)
+redir_out(src::AbstractCmd, dest::Redirectable) = CmdRedirect(src, dest, STDOUT_NO)
+redir_err(src::AbstractCmd, dest::Redirectable) = CmdRedirect(src, dest, STDERR_NO)
 
 # File redirects
-(|>)(src::AbstractCmd, dest::AbstractString) = CmdRedirect(src, FileRedirect(dest, false), STDOUT_NO)
-(|>)(src::AbstractString, dest::AbstractCmd) = CmdRedirect(dest, FileRedirect(src, false), STDIN_NO)
-(.>)(src::AbstractCmd, dest::AbstractString) = CmdRedirect(src, FileRedirect(dest, false), STDERR_NO)
-(>>)(src::AbstractCmd, dest::AbstractString) = CmdRedirect(src, FileRedirect(dest, true), STDOUT_NO)
-(.>>)(src::AbstractCmd, dest::AbstractString) = CmdRedirect(src, FileRedirect(dest, true), STDERR_NO)
+redir_out(src::AbstractCmd, dest::AbstractString) = CmdRedirect(src, FileRedirect(dest, false), STDOUT_NO)
+redir_out(src::AbstractString, dest::AbstractCmd) = CmdRedirect(dest, FileRedirect(src, false), STDIN_NO)
+redir_err(src::AbstractCmd, dest::AbstractString) = CmdRedirect(src, FileRedirect(dest, false), STDERR_NO)
+redir_out_append(src::AbstractCmd, dest::AbstractString) = CmdRedirect(src, FileRedirect(dest, true), STDOUT_NO)
+redir_err_append(src::AbstractCmd, dest::AbstractString) = CmdRedirect(src, FileRedirect(dest, true), STDERR_NO)
 
+function pipe(cmd::AbstractCmd; stdin=nothing, stdout=nothing, stderr=nothing, append::Bool=false)
+    if append && stdout === nothing && stderr === nothing
+        error("append set to true, but no output redirections specified")
+    end
+    if stdin !== nothing
+        cmd = redir_out(stdin, cmd)
+    end
+    if stdout !== nothing
+        cmd = append ? redir_out_append(cmd, stdout) : redir_out(cmd, stdout)
+    end
+    if stderr !== nothing
+        cmd = append ? redir_err_append(cmd, stderr) : redir_err(cmd, stderr)
+    end
+    return cmd
+end
+
+pipe(cmd::AbstractCmd, dest) = pipe(cmd, stdout=dest)
+pipe(src::Union(Redirectable,AbstractString), cmd::AbstractCmd) = pipe(cmd, stdin=src)
+
+pipe(a, b, c, d...) = pipe(pipe(a,b), c, d...)
 
 typealias RawOrBoxedHandle Union(UVHandle,UVStream,Redirectable,IOStream)
 typealias StdIOSet NTuple{3,RawOrBoxedHandle}